Rename `RepositoryInterceptor` and so on to `ChangeTrackingInterceptor`.
