+++ framework/src/Volo.Abp.Ddd.Domain/Volo/Abp/Domain/ChangeTracking/ChangeTrackingInterceptor.cs
@@ -1,21 +1,22 @@
 using System.Threading.Tasks;
 using Volo.Abp.DependencyInjection;
+using Volo.Abp.Domain.Repositories;
 using Volo.Abp.DynamicProxy;
 
-namespace Volo.Abp.Domain.Repositories;
+namespace Volo.Abp.Domain.ChangeTracking;
 
-public class RepositoryInterceptor : AbpInterceptor, ITransientDependency
+public class ChangeTrackingInterceptor : AbpInterceptor, ITransientDependency
 {
     private readonly IEntityChangeTrackingProvider _entityChangeTrackingProvider;
 
-    public RepositoryInterceptor(IEntityChangeTrackingProvider entityChangeTrackingProvider)
+    public ChangeTrackingInterceptor(IEntityChangeTrackingProvider entityChangeTrackingProvider)
     {
         _entityChangeTrackingProvider = entityChangeTrackingProvider;
     }
 
     public async override Task InterceptAsync(IAbpMethodInvocation invocation)
     {
-        if (!RepositoryHelper.IsEntityChangeTrackingMethod(invocation.Method, out var changeTrackingAttribute))
+        if (!ChangeTrackingHelper.IsEntityChangeTrackingMethod(invocation.Method, out var changeTrackingAttribute))
         {
             await invocation.ProceedAsync();
             return;
