+++ framework/src/Volo.Abp.Ddd.Domain/Volo/Abp/Domain/ChangeTracking/ChangeTrackingInterceptorRegistrar.cs
@@ -3,20 +3,20 @@ using System.Reflection;
 using Volo.Abp.DependencyInjection;
 using Volo.Abp.DynamicProxy;
 
-namespace Volo.Abp.Domain.Repositories;
+namespace Volo.Abp.Domain.ChangeTracking;
 
-public class RepositoryInterceptorRegistrar
+public class ChangeTrackingInterceptorRegistrar
 {
     public static void RegisterIfNeeded(IOnServiceRegistredContext context)
     {
         if (ShouldIntercept(context.ImplementationType))
         {
-            context.Interceptors.TryAdd<RepositoryInterceptor>();
+            context.Interceptors.TryAdd<ChangeTrackingInterceptor>();
         }
     }
 
     private static bool ShouldIntercept(Type type)
     {
-        return !DynamicProxyIgnoreTypes.Contains(type) && RepositoryHelper.IsEntityChangeTrackingType(type.GetTypeInfo());
+        return !DynamicProxyIgnoreTypes.Contains(type) && ChangeTrackingHelper.IsEntityChangeTrackingType(type.GetTypeInfo());
     }
 }
