+++ framework/src/Volo.Abp.Ddd.Domain/Volo/Abp/Domain/ChangeTracking/DisableEntityChangeTrackingAttribute.cs
@@ -1,11 +1,11 @@
 using System;
 
-namespace Volo.Abp.Domain.Repositories;
+namespace Volo.Abp.Domain.ChangeTracking;
 
 /// <summary>
 /// Ensures that the change tracking in enabled for the given method or class.
 /// </summary>
-[AttributeUsage(AttributeTargets.Method | AttributeTargets.Class | AttributeTargets.Interface)]
+[AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
 public class DisableEntityChangeTrackingAttribute : EntityChangeTrackingAttribute
 {
     public DisableEntityChangeTrackingAttribute()
